+++ app/services/draw_group_formation_starter.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
-# Service object to handle the activation of a draw. Checks to make sure the
+# Service object to handle the updating the draw status. Checks to make sure the
 # draw has the correct status, updates the status, and sends e-mail invitations
 # to students in the draw.
-class DrawActivator
+class DrawGroupFormationStarter
   include ActiveModel::Model
   include Callable
 
-  validate :draw_in_a_draft
+  validate :draw_in_intent_selection
   validate :draw_has_students
 
-  # Initialize a new DrawActivator
+  # Initialize a new DrawGroupFormationStarter
   #
   # @param draw [Draw] the draw in question
   # @param mailer [ActionMailer::Base] mailer class for sending invitation
@@ -21,29 +21,29 @@ class DrawActivator
     @college = College.current
   end
 
-  # Activate a Draw
+  # Update a Draw
   #
   # @return [Hash{Symbol=>ApplicationRecord,Hash}] A results hash with the
   #   message to set in the flash and either `nil` or the created object.
-  def activate
+  def start
     return error(self) unless valid?
-    draw.update!(status: 'intent_selection')
+    draw.update!(status: 'group_formation')
     send_emails
     success
   rescue ActiveRecord::RecordInvalid => e
     error(e)
   end
 
-  make_callable :activate
+  make_callable :start
 
   private
 
   attr_accessor :draw
   attr_reader :college, :mailer
 
-  def draw_in_a_draft
-    return if draw.draft?
-    errors.add(:draw, 'must be a draft.')
+  def draw_in_intent_selection
+    return if draw.intent_selection?
+    errors.add(:draw, 'must be in intent-selection.')
   end
 
   def draw_has_students
@@ -53,17 +53,17 @@ class DrawActivator
 
   def send_emails
     draw.students.each do |student|
-      mailer.draw_invitation(user: student, college: college).deliver_later
+      mailer.group_formation(user: student, college: college).deliver_later
     end
   end
 
   def success
-    { redirect_object: draw, msg: { notice: 'Draw successfully initiated.' } }
+    { redirect_object: draw, msg: { notice: 'Draw successfully updated.' } }
   end
 
   def error(error_obj)
     msg = ErrorHandler.format(error_object: error_obj)
     { redirect_object: nil,
-      msg: { error: "There was a problem initiating the draw:\n#{msg}" } }
+      msg: { error: "There was a problem updating the draw:\n#{msg}" } }
   end
 end