+++ spec/features/draws/draw_group_formation_start_spec.rb
@@ -3,7 +3,7 @@
 require 'rails_helper'
 
 RSpec.feature 'Draw activation' do
-  let(:draw) { create(:draw_with_members, status: 'draft') }
+  let(:draw) { create(:draw_with_members, status: 'intent_selection') }
 
   context 'as admin' do
     before do
@@ -13,15 +13,15 @@ RSpec.feature 'Draw activation' do
 
     it 'can be initiated' do
       visit draw_path(draw)
-      click_link('Begin intent selection phase')
+      click_link('Proceed to group formation phase')
       expect(page).to have_css('.flash-notice',
-                               text: 'Draw successfully initiated.')
+                               text: 'Draw successfully updated.')
     end
 
     it 'redirects on failure' do
-      draw = create(:draw, status: 'draft')
+      draw = create(:draw, status: 'intent_selection')
       visit draw_path(draw)
-      click_link('Begin intent selection phase')
+      click_link('Proceed to group formation phase')
       expect(page).to have_css('.flash-error')
     end
   end
@@ -31,7 +31,7 @@ RSpec.feature 'Draw activation' do
 
     it 'cannot see the start button' do
       visit draw_path(draw)
-      expect(page).not_to have_link('Begin intent selection phase')
+      expect(page).not_to have_link('Proceed to group formation phase')
     end
   end
 end